--- src/contexts/Authcontexts.tsx
@@ -1,30 +1,0 @@
-"use client";
-
-import UnauthorizedAccess from "@/components/user/UnauthorizedAccess/UnauthorizedAccess";
-import { getSessionId } from "@/utils/apiFunc";
-import { type ReactNode, createContext, useContext, useEffect, useState } from "react";
-
-// Contextオブジェクトを作成
-export const AuthContext = createContext<string | null>(null);
-
-type AuthProviderProps = {
-  children: ReactNode;
-};
-
-// Providerコンポーネント
-export const AuthProvider = ({ children }: AuthProviderProps) => {
-  const [token, setToken] = useState<string | null>(null);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    getSessionId(setToken, setLoading);
-  }, []);
-
-  if (loading) return <p>Loading...</p>;
-  if (!token) return <UnauthorizedAccess />;
-
-  return <AuthContext.Provider value={token}>{children}</AuthContext.Provider>;
-};
-
-// カスタムフック作成
-export const useAuth = () => useContext(AuthContext);
